feat: изменить float на double, убрать коммент

+++ src/Converter.java
@@ -1,9 +1,9 @@
 public class Converter {
-    float kiloCalories;
-    float distance;
+    double kiloCalories;
+    double distance;
 
-    Converter(float steps) {
-        distance = (float) (steps * 0.00075);
-        kiloCalories = (float) (steps * 0.05);
+    Converter(double steps) {
+        distance = steps * 0.00075;
+        kiloCalories = steps * 0.05;
     }
 }
